# openapi.yaml
openapi: 3.0.3
info:
  title: MySelfServe API
  version: 0.1.0
  description: |
    BC Government Income Assistance Self-Service Portal API.

    MySS 2.0 provides self-service capabilities for BC income assistance clients
    to manage service requests, monthly reports, payments, employment plans,
    messages, and account settings. Ministry workers can access client data
    through an admin support view.

    ## Authentication

    Most endpoints require a JWT Bearer token issued via BCeID/IDIR through
    Keycloak OIDC. The token is signed with HS256 and contains claims for
    `sub` (user_id), `role`, `bceid_guid`, and `idir_username`.

    ## Roles

    - **CLIENT** — BCeID-authenticated income assistance clients
    - **WORKER** — IDIR-authenticated ministry staff (includes ADMIN)
    - **ADMIN** — System administrators (inherits WORKER permissions)

    ## Error Handling

    Errors from the upstream Siebel/ICM system are mapped to standard HTTP
    status codes with structured JSON error bodies. A circuit breaker protects
    the application during ICM outages (5 failures → 30s open window → 503).
  contact:
    name: MySS Development Team
  license:
    name: BC Government
    url: https://www2.gov.bc.ca/
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://myss-api-dev.apps.silver.devops.gov.bc.ca
    description: Development (OpenShift)
  - url: https://myss-api-test.apps.silver.devops.gov.bc.ca
...
security:
  - BearerAuth: []
paths:
...
  /monthly-reports:
    get:
      operationId: listMonthlyReports
      tags: [Monthly Reports]
      summary: List monthly reports
      parameters:
        - name: days_ago
          in: query
          description: Number of days of history to retrieve
          schema:
            type: integer
            minimum: 1
            default: 365
      responses:
        "200":
          description: List of monthly reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SD81ListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    post:
      operationId: createMonthlyReport
      tags: [Monthly Reports]
      summary: Create a new monthly report draft
      responses:
        "201":
          description: Monthly report draft created
          content:
            application/json:
              schema:
                type: object
                properties:
                  sd81_id:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /monthly-reports/current-period:
    get:
      operationId: getCurrentReportPeriod
      tags: [Monthly Reports]
      summary: Get the current reporting period schedule
      parameters:
        - name: case_number
          in: query
          schema:
            type: string
            default: default
      responses:
        "200":
          description: Current reporting period
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChequeScheduleWindow"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /monthly-reports/{sd81_id}/answers:
    get:
      operationId: getMonthlyReportAnswers
      tags: [Monthly Reports]
      summary: Get saved answers for a monthly report
      parameters:
        - name: sd81_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report answers
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"

    put:
      operationId: saveMonthlyReportAnswers
      tags: [Monthly Reports]
      summary: Save answers for a monthly report
      parameters:
        - name: sd81_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Answers saved
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
  /monthly-reports/{sd81_id}/submit:
    post:
      operationId: submitMonthlyReport
      tags: [Monthly Reports]
      summary: Submit a monthly report
      parameters:
        - name: sd81_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SD81SubmitRequest"
      responses:
        "200":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SD81SubmitResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Invalid PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Submission period closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /monthly-reports/{sd81_id}/restart:
    post:
      operationId: restartMonthlyReport
      tags: [Monthly Reports]
      summary: Restart a monthly report
      parameters:
        - name: sd81_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report restarted
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
  /monthly-reports/{sd81_id}/pdf:
    get:
      operationId: downloadMonthlyReportPdf
      tags: [Monthly Reports]
      summary: Download a monthly report as PDF
      parameters:
        - name: sd81_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: PDF document
          headers:
            Content-Disposition:
              schema:
                type: string
                example: "inline; filename=SD81_12345.pdf"
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ──────────────────────────────────────────────
  # Notifications & Messages
  # ──────────────────────────────────────────────
  /notifications/banners:
    get:
      operationId: getBanners
      tags: [Notifications]
      summary: Get active banner notifications
      responses:
        "200":
          description: Active banners
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BannerListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /messages:
    get:
      operationId: listMessages
      tags: [Messages]
      summary: List inbox messages
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        "200":
          description: Paginated inbox
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InboxListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /messages/{msg_id}:
    get:
      operationId: getMessage
      tags: [Messages]
      summary: Get message details
      description: Returns the full message body and attachments. Marks the message as read in the background.
      parameters:
        - name: msg_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Message details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageDetail"
        "401":
          $ref: "#/components/responses/Unauthorized"

    delete:
      operationId: deleteMessage
      tags: [Messages]
      summary: Delete a message
      parameters:
        - name: msg_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Message deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/{msg_id}/reply:
    post:
      operationId: replyToMessage
      tags: [Messages]
      summary: Reply to a message
      parameters:
        - name: msg_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplyRequest"
      responses:
        "200":
          description: Reply sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Reply not allowed for this message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /messages/{msg_id}/sign:
    post:
      operationId: signMessage
      tags: [Messages]
      summary: Sign a message with PIN
      parameters:
        - name: msg_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignRequest"
      responses:
        "200":
          description: Message signed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: signed
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Invalid PIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # ──────────────────────────────────────────────
  # Payment
  # ──────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token signed with HS256. Claims: `sub` (user_id), `role` (CLIENT/WORKER/ADMIN),
        `bceid_guid` (clients), `idir_username` (workers/admins), `exp` (expiry).
    WebhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: HMAC secret for internal AV scanner webhook
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    ServiceUnavailable:
      description: Upstream service (ICM/Siebel) unavailable — circuit breaker open
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Human-readable error message
      required:
        - detail
...
    SD81Status:
      type: string
      enum: [PRT, SUB, RST, RES, PND]
      description: |
        Monthly report status codes:
        - PRT: Partial (in progress)
        - SUB: Submitted
        - RST: Restarted
        - RES: Resubmitted
        - PND: Pending Documents
    ICMMessageType:
      type: string
      enum:
        - HR0081
        - HR0081Restart
        - HR0081StreamlinedRestart
        - "HR0081 Pending Documents"
        - FormSubmission
        - General
...
    ChequeScheduleWindow:
      type: object
      required:
        - benefit_month
        - income_date
        - cheque_issue_date
        - period_close_date
      properties:
        benefit_month:
          type: string
          format: date
        income_date:
          type: string
          format: date
        cheque_issue_date:
          type: string
          format: date
        period_close_date:
          type: string
          format: date-time
          description: Friday 4 PM UTC
    SD81Summary:
      type: object
      required:
        - sd81_id
        - benefit_month
        - status
      properties:
        sd81_id:
          type: string
        benefit_month:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/SD81Status"
        submitted_at:
          type: string
          format: date-time
          nullable: true
    SD81ListResponse:
      type: object
      required:
        - reports
        - total
      properties:
        reports:
          type: array
          items:
            $ref: "#/components/schemas/SD81Summary"
        total:
          type: integer
    SD81SubmitRequest:
      type: object
      required:
        - pin
      properties:
        pin:
          type: string
        spouse_pin:
          type: string
          nullable: true
    SD81SubmitResponse:
      type: object
      required:
        - sd81_id
        - status
        - submitted_at
      properties:
        sd81_id:
          type: string
        status:
          $ref: "#/components/schemas/SD81Status"
        submitted_at:
          type: string
          format: date-time

    # ── Notifications & Messages ────────────────
    BannerNotification:
      type: object
      required:
        - notification_id
        - body
        - start_date
        - end_date
      properties:
        notification_id:
          type: string
        body:
          type: string
          maxLength: 150
        start_date:
          type: string
          format: date-time
        end_date:
          type: string
          format: date-time
    BannerListResponse:
      type: object
      required:
        - banners
      properties:
        banners:
          type: array
          items:
            $ref: "#/components/schemas/BannerNotification"
    MessageSummary:
      type: object
      required:
        - message_id
        - subject
        - sent_date
        - is_read
        - can_reply
        - message_type
      properties:
        message_id:
          type: string
        subject:
          type: string
        sent_date:
          type: string
          format: date-time
        is_read:
          type: boolean
        can_reply:
          type: boolean
        message_type:
          $ref: "#/components/schemas/ICMMessageType"
    MessageDetail:
      type: object
      required:
        - message_id
        - subject
        - sent_date
        - is_read
        - can_reply
        - message_type
        - body
        - attachments
      properties:
        message_id:
          type: string
        subject:
          type: string
        sent_date:
          type: string
          format: date-time
        is_read:
          type: boolean
        can_reply:
          type: boolean
        message_type:
          $ref: "#/components/schemas/ICMMessageType"
        body:
          type: string
        attachments:
          type: array
          items:
            type: string
    InboxListResponse:
      type: object
      required:
        - messages
        - total
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/MessageSummary"
        total:
          type: integer
    ReplyRequest:
      type: object
      required:
        - body
      properties:
        body:
          type: string
          maxLength: 4000
        attachment_ids:
          type: array
          items:
            type: string
          default: []
    ReplyResponse:
      type: object
      required:
        - status
        - sent_at
      properties:
        status:
          type: string
        sent_at:
          type: string
          format: date-time
    SignRequest:
      type: object
      required:
        - pin
      properties:
        pin:
          type: string
          minLength: 1

    # ── Payment ─────────────────────────────────
...